backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.1
asyncpg>=0.30.0
SQLAlchemy>=2.0.41
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.1
redis
fastapi-cache2[redis]>=0.2.2
numpy
pandas>=2.0.0
scipy
requests
greenlet
torch
